Extract class SliceFilter, BalancedSample and SampleMetaData from loader.utils
