+++ dev/metadata_config.py
@@ -6,9 +6,8 @@ from torchvision import transforms as torch_transforms
 
 from ivadomed.loader.bids_dataset import BidsDataset
 from ivadomed import config_manager as imed_config_manager
-from ivadomed.loader import utils as imed_loader_utils
+from ivadomed.loader.slice_filter import SliceFilter
 from ivadomed import transforms as imed_transforms
-from ivadomed import utils as imed_utils
 
 metadata_type = ['FlipAngle', 'EchoTime', 'RepetitionTime']
 
@@ -29,7 +28,7 @@ def run_main(context):
                              contrast_lst=context["contrast_train_validation"]
                              if subset != "test" else context["contrast_test"],
                              transform=no_transform,
-                             slice_filter_fn=imed_loader_utils.SliceFilter())
+                             slice_filter_fn=SliceFilter())
 
             for m in metadata_type:
                 if m in metadata_dct:
