+++ ivadomed/loader/sample_meta_data.py
@@ -1,0 +1,41 @@
+class SampleMetadata(object):
+    """Metadata class to help update, get and set metadata values.
+
+    Args:
+        d (dict): Initial metadata.
+
+    Attributes:
+        metadata (dict): Image metadata.
+    """
+
+    def __init__(self, d=None):
+        self.metadata = {} or d
+
+    def __setitem__(self, key, value):
+        self.metadata[key] = value
+
+    def __getitem__(self, key):
+        return self.metadata[key]
+
+    def __contains__(self, key):
+        return key in self.metadata
+
+    def items(self):
+        return self.metadata.items()
+
+    def _update(self, ref, list_keys):
+        """Update metadata keys with a reference metadata.
+
+        A given list of metadata keys will be changed and given the values of the reference
+        metadata.
+
+        Args:
+            ref (SampleMetadata): Reference metadata object.
+            list_keys (list): List of keys that need to be updated.
+        """
+        for k in list_keys:
+            if (k not in self.metadata.keys() or not bool(self.metadata[k])) and k in ref.metadata.keys():
+                self.metadata[k] = ref.metadata[k]
+
+    def keys(self):
+        return self.metadata.keys()
