+++ testing/unit_tests/test_sampler.py
@@ -6,6 +6,7 @@ from torch.utils.data import DataLoader
 from ivadomed.loader.bids_dataframe import BidsDataframe
 from ivadomed import utils as imed_utils
 from ivadomed.loader import utils as imed_loader_utils, loader as imed_loader
+from ivadomed.loader.balanced_sampler import BalancedSampler
 from testing.unit_tests.t_utils import create_tmp_dir,  __data_testing_dir__, __tmp_dir__, download_data_testing_test_files
 from testing.common_testing_util import remove_tmp_dir
 
@@ -86,7 +87,7 @@ def test_sampler(download_data_testing_test_files, transforms_dict, train_lst, t
 
     print('\nLoading with sampling')
     train_loader_balanced = DataLoader(ds_train, batch_size=BATCH_SIZE,
-                                       sampler=imed_loader_utils.BalancedSampler(ds_train),
+                                       sampler=BalancedSampler(ds_train),
                                        shuffle=False, pin_memory=True,
                                        collate_fn=imed_loader_utils.imed_collate,
                                        num_workers=0)
